[modified] BE pet level calculation to better emulate live

+++ MMOCoreORB/src/server/zone/managers/creature/DnaManager.h
@@ -33,9 +33,11 @@ protected:
 public:
 	DnaManager();
 	virtual ~DnaManager();
+
 	void loadSampleData();
 	int generateXp(int creatureLevel);
 	static int addQualityTemplate(lua_State* L);
+
 	enum {
 		FORTITUDE     = 1,
 		ENDURANCE     = 2,
@@ -48,6 +50,7 @@ public:
 		POWER         = 9,
 		FIERCENESS    = 10
 	};
+
 	enum {
 		HIT_LEVEL = 1,
 		HAM_LEVEL = 2,
@@ -55,14 +58,17 @@ public:
 		ARM_LEVEL = 4,
 		REG_LEVEL = 5
 	};
+
 	void generateSample(Creature* creature, CreatureObject* player, int quality);
 	void generationalSample(PetDeed* deed, CreatureObject* player, int quality);
-	int levelForScore(int type, float value);
+	float levelForScore(int type, float value);
 	float valueForLevel(int type, int level);
+
 protected:
 	int reduceByPercent(int source, int percent) {
 		float reduceBy = (100.0 - (float)percent) / 100.0;
-		uint32 newValue = (int)( ((float)source) * reduceBy );
+		uint32 newValue = (int)( ((float)source) * reduceBy);
+
 		if (newValue < 1)
 			newValue = 1;
 		if (newValue > 999)
